Remove kas_core::draw::images path-deduplication module
